Move Patch, AtomicPut and MasterService tests to test/integration.
